fix: ignore misc wipe always / breaks OTA* & Android factory reset
This commit will change the default behavior for the use of /misc in TWRP.

/misc is where in general bootloader args went e.g when going through
the factory reset screen or in Android when starting the factory reset.

The G4 has no hardware key combo to bring up the recovery so to be
able to boot into TWRP the factory reset screen is used as a trampoline here.

The factory reset screen writes into the /misc partition to enforce the wipe
so in TWRP /misc was not linked to the real misc blockdevice but to /grow as
all newer custom ROMs doing the same. Due to this nothing happens when
the hw key combo is used and just booting TWRP.
A factory reset initiated by a custom ROM (or OTA updates) will write the
bootloader arg to /grow and as TWRP had read /grow (interpreted as /misc) it
was able to act accordingly (means: reset or applying an update).

This all goes well as long as... when using e.g. fastboot erase.
Until now that one was unknown to me: fastboot erase also writes into /misc
when doing this:
fastboot erase userdata (not fastboot format).
This does not *format* the data and the following should describe what
can happen and the new way I go:

**************************
Scenarios and the new behavior:

A) previous official and unofficial custom-mm and custom-n builds:
TW_IGNORE_MISC_WIPE_DATA is *disabled* and /misc -> /grow
fastboot erase will write into /misc and TWRP will bootloop without an
option to bring up TWRP!
Reason: the real misc blockdevice gets never cleaned as the link to /misc
is /grow.
It seems that aboot may reads the misc blockdevice and directly boots into
recovery when needed and as that never get cleaned the bootloop occurs
with an unknown error 150 (which means TWRP tries a wipe and not a format
and so failing)

B) previous unofficial stock builds:
TW_IGNORE_MISC_WIPE_DATA is *enabled* and /misc -> /grow
fastboot erase will write into /misc and TWRP will bootloop but is usable.
Reason: This occurs e.g. on STOCK based ROMs as they mount /misc as
it should be but TWRP reads and cleans only /grow.

C) never released anywhere (for a reason):
TW_IGNORE_MISC_WIPE_DATA is *disabled* and /misc -> /misc
fastboot, factory reset and OTA for STOCK will work fine and even no
bootloop will occur but! we will loose the hardware key combo to open
TWRP!! so absolutely nothing we want.

new default behaviour with this commit:
D) TW_IGNORE_MISC_WIPE_DATA is *enabled* and /misc -> /misc
fastboot erase will not break anything (you need to choose format in TWRP
manually though!)
factory reset screen via hardware combo will just open TWRP
factory reset in Android and OTA's (of ROMs using /grow as /misc) will
just open TWRP without doing anything
OTA's (of ROMs using /misc as /misc) will work

This way no bootloop would ever occur (hopefully). If a bootloop still occurs
the ROM has to be updated to use real /misc blockdevice.

Factory reset has to be done in TWRP now always. There is no way around it
but thats better as having a bootloop imho.

* To enable OTA from within Android again the fstab of the ROM has to be
changed to use the real /misc blockdevice.

To be backwards compatible with previous TWRP released /rmisc will stay
in the TWRP fstab as well (no need to do this in any ROM).

(cherry picked from commit d8053c20a97607d709033854b7ad99bba45dd97b)

# Conflicts:
#	recovery.fstab