fix: decryption not working
since I removed the firmware mount as vfat qseecomd was
not able to find required firmware files like keymaster etc.

The reason for the previously removal of mounting firmware
as vfat was trouble in restoring it.

The way it is implemented now will ensure that a backup will
be done as image while still be possible to mount it as vfat
when needed

Change-Id: I9b54f95dc8331ef342e3ce3ce9782ae26c680d6f

(cherry picked from commit 0551a64a7f97ab85e86e1516f593b72ad3182154)