g4: fstab: fix encrypt partition backup/restore
doing a backup for that partition must be done as emmc
partition which results in a dump image.
defining it as emmc will not allow it to wipe though as
mke2fs does not know what to do on that "fake" emmc fs type
used by TWRP.

the fix is to define it twice (emmc + a valid fs type) and set the
flags properly for each situation!

flags:
- `removable` as that partition should not be auto mounted
- `wipeduringfactoryreset` to auto wipe when factory reset
- `wipeingui` to allow a manual wipe
- `backup=1|0` to allow backup as raw image while denying the other

note:
in this particular case formatting as vfat is ok as qseecomd
does not care about how it is formatted (as it writes raw to
that partition). ofc after encrypting that partition is likely
not mountable anymore but that does not matter.

Background:
https://github.com/Suicide-Squirrel/issues_pie/issues/52

Signed-off-by: steadfasterX <steadfasterX@gmail.com>
(cherry picked from commit c0c067e771141bab1df1cc15402dfd3e0ac74600)